[Presentation] Moved the shared dialog widgets to its own folder
