+++ lib/presentation/notifications/widgets/forms/notification_realm_currency_form.dart
@@ -7,7 +7,7 @@ import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/generated/l10n.dart';
 import 'package:genshindb/presentation/shared/dropdown_button_with_title.dart';
 import 'package:genshindb/presentation/shared/extensions/i18n_extensions.dart';
-import 'package:genshindb/presentation/shared/number_picker_dialog.dart';
+import 'package:genshindb/presentation/shared/dialogs/number_picker_dialog.dart';
 
 import 'notification_circle_item.dart';
 import 'notification_dropdown_type.dart';
